pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "douyin-xiaohongshu-monitor"
version = "0.1.0"
description = "Douyin and Xiaohongshu monitoring service."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.13,<2.0",
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.4,<3.0",
  "redis>=5.0,<6.0",
  "apscheduler>=3.10,<4.0",
  "httpx>=0.27,<1.0",
  "playwright>=1.52,<2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
